refact: move dtos from `dto` to `api` package
