+++ src/main/java/com/peerloop/peerloopapp/domain/auth/api/response/AuthResponse.java
@@ -1,4 +1,4 @@
-package com.peerloop.peerloopapp.domain.auth.dto.response;
+package com.peerloop.peerloopapp.domain.auth.api.response;
 
 public record AuthResponse(
         String accessToken,
