+++ src/main/java/com/peerloop/peerloopapp/domain/member/api/request/SignUpRequest.java
@@ -1,4 +1,4 @@
-package com.peerloop.peerloopapp.domain.member.dto.request;
+package com.peerloop.peerloopapp.domain.member.api.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
